v0.2 (Timer added)
First implementation idea of a timer instead of k_msleep, based on the examples on the zephyr kernel timer documentation. As of right now, build works without errors or warnings, so i take it as if its a correct implementation. Also, some methods and bits of code which may need explanation dont have it, ill add it later.