v0.5 (LEDs for time display working?)
Suposedly this should be the first part of the technical test, in which the LEDs display the time the system has been active. I´ve used a timer which every minute sends the workload with the time updating and leds switching to the system queue and goes back to waiting an other minute. Ive chosen this way because i believe that the workload i send to the system queue doesnt interfere with the actual timer (by what I´ve understood reading the documentation of the timer). The next changes will be implementing the button reset functionality, after that i will re-check everything is suposed to be as intended and make sure the code follows some kind of standard. Like in previous commits, the code as of right now builds without errors or warnings, so i believe most part of it is correct, at least the parts involving interactions with the hardware of the board.