v0.6 (Initial button functionality added)
Added the button functionality based mainly on the Button example given in the documentation, our board already has a alias to relate sw0 and button0 in the device tree, so i didnt really need to change much from the example. Later ill check if this would really work as intended, and add comments if needed. Like in previous commits, the code builds successfully, without any error or warning, so at least the hardware connections should be correct.