Multithread with one semaphore (Previous commit wont work)
This would be the implementation if we used one semaphore, it feels wrong having a task only post a semaphore and the other one only get the semaphore, but with one semaphore this would be the implementation. Also clarify that the previous commit would not work because i just realized that i didnt put a while loop in the leds_thread, so it would only display the time once. In this commit its fixed.